+++ adapters/AppLovin/AppLovinAdapter/GADMAdapterAppLovinRewardBasedVideoAd.swift
@@ -1,4 +1,4 @@
-// Copyright 2018 Google LLC
+// Copyright 2017 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,19 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#import "GADMAdapterAppLovin.h"
-#import "AppLovinAdapter-Swift.h"
-#import "GADMediationAdapterAppLovin.h"
+import Foundation
+import GoogleMobileAds
 
-/// The adapter class that Google Serving refers to for loading AppLovin Banner and Interstitial
-/// Waterfall ads.
-///
-/// Using the function mainAdapterClass, Google Mobile Ads SDK infers that the actual
-/// implementation is in GADMediationAdapterAppLovin.
-@implementation GADMAdapterAppLovin
+@MainActor
+@objc(GADMAdapterAppLovinRewardBasedVideoAd)
+public final class GADMAdapterAppLovinRewardBasedVideoAd: NSObject {
 
-+ (nonnull Class<GADMediationAdapter>)mainAdapterClass {
-    return [GADMediationAdapterAppLovin class];
+  /// TODO(Google): Remove this class once Google's server points to GADMediationAdapterAppLovin
+  /// directly to ask for a rewarded ad on waterfall mediation requests.
+  @objc public nonisolated static func mainAdapterClass() -> any MediationAdapter.Type {
+    return GADMediationAdapterAppLovin.self
+  }
 }
-
-@end